Add files via upload
A finished draft of a mini game within the larger game. Functional and essentially complete, but the code could do with some cleaning up and work will need to be done to integrate it into the completed game. A card game of (mostly) chance with a little bit of strategy, but you never know which way things will go. 
Dictionaries and lists used to deal the cards and build hands for the user and 3 virtual players